# shared-schema/queueserver_service.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bluesky HTTP Server",
    "description": "REST + WebSocket surface of the Bluesky Queue Server. Drives an experiment from\nthree angles: **queue** (plans and instructions waiting to run), **Run Engine**\n(control over the currently-running plan), and **worker environment** (startup\nscripts, plans, devices, and tasks inside the worker process).\n\nAll endpoints live under `/api`. Clients authenticate with a bearer token\n(access or API key); anonymous access, when enabled, is scoped to `read:status`\nonly. Most endpoints return `{\"success\": bool, \"msg\": str, ...}` envelopes \u2014\ninspect `success` on each response.\n\n### WebSocket endpoints (not listed below)\n\nThree push streams are available; they do not render in Swagger UI because\nOpenAPI has no WebSocket concept:\n\n- `GET /api/status/ws` \u2014 pushes the `/api/status` payload as it changes\n  (scope `read:monitor`).\n- `GET /api/info/ws` \u2014 pushes system-info snapshots (scope `read:monitor`).\n- `GET /api/console_output/ws` \u2014 pushes captured worker stdout/stderr\n  (scope `read:console`).\n\nClients authenticate WebSockets by passing the bearer token via a cookie or\nthe `Authorization` header on the upgrade request.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get RE Manager status",
        "description": "Returns a status snapshot of RE Manager \u2014 manager state, environment state, the currently running item (if any), worker process status, queue/history counts, plus the UIDs clients use for change detection when polling. Required scope: `read:status`.",
        "operationId": "status_handler_api_status_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "read:status"
            ]
          },
          {
            "APIKeyQuery": [
              "read:status"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "read:status"
            ]
          },
          {
            "APIKeyCookie": [
              "read:status"
            ]
          }
        ]
      }
    },
...
    "/api/profile_collection/status": {
      "get": {
        "tags": [
          "Profile Collection"
        ],
        "summary": "Status of the on-disk profile collection",
        "description": "Returns the current commit, branch, dirty flag, and ahead/behind counts relative to the configured upstream. Cheap; safe to poll from the UI for the Profile Collection header badge. Required scope: `read:status`.",
        "operationId": "profile_collection_status_handler_api_profile_collection_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "read:status"
            ]
          },
          {
            "APIKeyQuery": [
              "read:status"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "read:status"
            ]
          },
          {
            "APIKeyCookie": [
              "read:status"
            ]
          }
        ]
      }
    },
    "/api/profile_collection/pull": {
      "post": {
        "tags": [
          "Profile Collection"
        ],
        "summary": "Fast-forward the profile collection from upstream",
        "description": "Runs `git fetch` then `git merge --ff-only @{upstream}` in the profile-collection directory. Rejected (409) when the working tree is dirty \u2014 operator must commit or revert on-host edits first. Rejected (409) when upstream has diverged (non-fast-forward) \u2014 operator must resolve on-host. Returns commit_before, commit_after, the diff file list, and a pixi_toml_changed flag the UI uses to decide whether a hard service restart is needed before reload. Required scope: `write:manager:control`.",
        "operationId": "profile_collection_pull_handler_api_profile_collection_pull_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilePullResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyQuery": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyCookie": [
              "write:manager:control"
            ]
          }
        ]
      }
    },
    "/api/profile_collection/reload": {
      "post": {
        "tags": [
          "Profile Collection"
        ],
        "summary": "Pull and recycle the worker (one-shot for the UI)",
        "description": "Wrapper that runs `pull`, then \u2014 if an environment is open \u2014 closes and reopens it so the new profile is loaded into the worker. Does NOT mutate the device registry; the UI should follow up with `GET /api/devices/diff_against_profile` and (on operator confirm) `POST /api/devices/sync_from_profile`. Returns 409 when the working tree is dirty or when pixi.toml changed and an environment is currently open (pixi.toml changes require a service-level restart \u2014 see ophyd-service#61). Required scope: `write:manager:control`.",
        "operationId": "profile_collection_reload_handler_api_profile_collection_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileReloadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyQuery": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyCookie": [
              "write:manager:control"
            ]
          }
        ]
      }
    },
    "/api/devices/diff_against_profile": {
      "get": {
        "tags": [
          "Profile Collection"
        ],
        "summary": "Diff worker-introspected devices against the registry (stub)",
        "description": "Will return added/removed/modified device specs vs configuration_service. Non-destructive. Backed by a compute_diff helper that this PR cannot land yet because the split inside manager/config_service.py collides with #59. Returns 501 until the follow-up commit on this branch lands. See ophyd-service#61.",
        "operationId": "devices_diff_against_profile_stub_api_devices_diff_against_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "read:status"
            ]
          },
          {
            "APIKeyQuery": [
              "read:status"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "read:status"
            ]
          },
          {
            "APIKeyCookie": [
              "read:status"
            ]
          }
        ]
      }
    },
    "/api/devices/sync_from_profile": {
      "post": {
        "tags": [
          "Profile Collection"
        ],
        "summary": "Apply a selected/all device diff to the registry (stub)",
        "description": "Will accept {strategy: 'all'|'additions_only'|'selected', devices?: [...]} and upsert/remove accordingly. Returns 501 until the follow-up commit on this branch lands. See ophyd-service#61.",
        "operationId": "devices_sync_from_profile_stub_api_devices_sync_from_profile_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyQuery": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyAuthorizationHeader": [
              "write:manager:control"
            ]
          },
          {
            "APIKeyCookie": [
              "write:manager:control"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LockState": {
        "properties": {
          "environment": {
            "type": "boolean",
            "title": "Environment"
          },
          "queue": {
            "type": "boolean",
            "title": "Queue"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "environment",
          "queue"
        ],
        "title": "LockState"
      },
...
      "PlanQueueMode": {
        "properties": {
          "loop": {
            "type": "boolean",
            "title": "Loop"
          },
          "ignore_failures": {
            "type": "boolean",
            "title": "Ignore Failures"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "loop",
          "ignore_failures"
        ],
        "title": "PlanQueueMode"
      },
...
      "ProfilePullResponse": {
        "properties": {
          "commit_before": {
            "type": "string",
            "title": "Commit Before",
            "description": "HEAD SHA before the pull."
          },
          "commit_after": {
            "type": "string",
            "title": "Commit After",
            "description": "HEAD SHA after the pull."
          },
          "files_changed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files Changed",
            "description": "Files that differ between commit_before and commit_after. Empty list when the pull was a no-op."
          },
          "pixi_toml_changed": {
            "type": "boolean",
            "title": "Pixi Toml Changed",
            "description": "True when pixi.toml is in files_changed. The HTTP layer uses this to surface requires_hard_restart to the UI; the endpoint itself does not 409 because the operator may have intentionally pulled without an open environment."
          }
        },
        "type": "object",
        "required": [
          "commit_before",
          "commit_after",
          "pixi_toml_changed"
        ],
        "title": "ProfilePullResponse",
        "description": "Outcome of a successful fast-forward pull."
      },
      "ProfileReloadResponse": {
        "properties": {
          "pull": {
            "$ref": "#/components/schemas/ProfilePullResponse"
          },
          "environment_recycled": {
            "type": "boolean",
            "title": "Environment Recycled",
            "description": "True if an existing environment was closed and reopened. False when no environment was open at call time \u2014 the operator must call /api/environment/open separately to trigger device introspection."
          }
        },
        "type": "object",
        "required": [
          "pull",
          "environment_recycled"
        ],
        "title": "ProfileReloadResponse",
        "description": "Outcome of the wrapper that pulls + recycles the worker."
      },
      "ProfileStatusResponse": {
        "properties": {
          "profile_dir": {
            "type": "string",
            "title": "Profile Dir",
            "description": "Absolute path on the server."
          },
          "commit": {
            "type": "string",
            "title": "Commit",
            "description": "HEAD commit SHA (40-char hex)."
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch",
            "description": "Current branch name, or null if HEAD is detached."
          },
          "is_dirty": {
            "type": "boolean",
            "title": "Is Dirty",
            "description": "True when the working tree has uncommitted changes (staged, unstaged, or untracked)."
          },
          "ahead": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ahead",
            "description": "Commits ahead of the configured upstream, or null if no upstream is set."
          },
          "behind": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behind",
            "description": "Commits behind the configured upstream, or null if no upstream is set."
          }
        },
        "type": "object",
        "required": [
          "profile_dir",
          "commit",
          "is_dirty"
        ],
        "title": "ProfileStatusResponse",
        "description": "Status of the on-disk profile collection."
      },
...
      "StatusResponse": {
        "properties": {
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "items_in_queue": {
            "type": "integer",
            "title": "Items In Queue"
          },
          "items_in_history": {
            "type": "integer",
            "title": "Items In History"
          },
          "running_item_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Running Item Uid"
          },
          "manager_state": {
            "type": "string",
            "title": "Manager State"
          },
          "queue_stop_pending": {
            "type": "boolean",
            "title": "Queue Stop Pending"
          },
          "queue_autostart_enabled": {
            "type": "boolean",
            "title": "Queue Autostart Enabled"
          },
          "worker_environment_exists": {
            "type": "boolean",
            "title": "Worker Environment Exists"
          },
          "worker_environment_state": {
            "type": "string",
            "title": "Worker Environment State"
          },
          "worker_background_tasks": {
            "type": "integer",
            "title": "Worker Background Tasks"
          },
          "re_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Re State"
          },
          "ip_kernel_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Kernel State"
          },
          "ip_kernel_captured": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Kernel Captured"
          },
          "pause_pending": {
            "type": "boolean",
            "title": "Pause Pending"
          },
          "status_uid": {
            "type": "string",
            "title": "Status Uid"
          },
          "run_list_uid": {
            "type": "string",
            "title": "Run List Uid"
          },
          "plan_queue_uid": {
            "type": "string",
            "title": "Plan Queue Uid"
          },
          "plan_history_uid": {
            "type": "string",
            "title": "Plan History Uid"
          },
          "devices_existing_uid": {
            "type": "string",
            "title": "Devices Existing Uid"
          },
          "plans_existing_uid": {
            "type": "string",
            "title": "Plans Existing Uid"
          },
          "devices_allowed_uid": {
            "type": "string",
            "title": "Devices Allowed Uid"
          },
          "plans_allowed_uid": {
            "type": "string",
            "title": "Plans Allowed Uid"
          },
          "plan_queue_mode": {
            "$ref": "#/components/schemas/PlanQueueMode"
          },
          "task_results_uid": {
            "type": "string",
            "title": "Task Results Uid"
          },
          "lock_info_uid": {
            "type": "string",
            "title": "Lock Info Uid"
          },
          "lock": {
            "$ref": "#/components/schemas/LockState"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "msg",
          "time",
          "items_in_queue",
          "items_in_history",
          "manager_state",
          "queue_stop_pending",
          "queue_autostart_enabled",
          "worker_environment_exists",
          "worker_environment_state",
          "worker_background_tasks",
          "pause_pending",
          "status_uid",
          "run_list_uid",
          "plan_queue_uid",
          "plan_history_uid",
          "devices_existing_uid",
          "plans_existing_uid",
          "devices_allowed_uid",
          "plans_allowed_uid",
          "plan_queue_mode",
          "task_results_uid",
          "lock_info_uid",
          "lock"
        ],
        "title": "StatusResponse",
        "description": "RE Manager status snapshot (``manager.py`` ``_status_update``). All keys always present."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "PLACEHOLDER",
            "refreshUrl": "token/refresh"
          }
        }
      },
      "APIKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key"
      },
      "APIKeyAuthorizationHeader": {
        "type": "apiKey",
        "description": "Prefix value with 'Apikey ' as in, 'Apikey SECRET'",
        "in": "header",
        "name": "Authorization"
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bluesky_httpserver_api_key"
      }
    }
  },
...
}